Complete issue step 1: Examine git_helper.py and git_operations.py to understand their responsibilities and plan the merge.
