+++ src/auto_slopp/workers/vikunja_task_source.py
@@ -8,7 +8,7 @@ import logging
 from pathlib import Path
 from typing import List
 
-from auto_slopp.utils.git_helper import commit
+from auto_slopp.utils.git_operations import commit
 from auto_slopp.utils.git_operations import sanitize_branch_name
 from auto_slopp.utils.vikunja_operations import (
     analyze_task,
